Create stacking
The program performs stacking ensemble learning operation on the Caravan dataset and successfully improves the model accuracy. In the first level applies 4 classifiers and calculates the accuracy score of these classifiers. In the second level combines the prediction of all these classifiers to form a new data frame and trains the machine using logistic regression. The final prediction is of 94% accuracy